requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Migrate>=4.0.5
Flask-CORS>=4.0.0
Flask-Mail>=0.9.1
SQLAlchemy>=2.0.36
GeoAlchemy2>=0.14.2
psycopg2-binary>=2.9.9
WTForms>=3.1.1
Werkzeug>=3.0.1
python-dotenv>=1.0.0
gunicorn>=21.0.0
Pillow>=10.2.0
barcode>=0.15.1
reportlab>=4.0.0
APScheduler>=3.10.0

# Performance & Caching
Flask-Caching>=2.1.0
redis>=5.0.0

# Rate Limiting
Flask-Limiter>=3.5.0

# Production Server (Linux/Mac)
# gunicorn is included above
# For Windows, use:
waitress>=3.0.0